more work

+++ src/libs/database/mirrorbraindb_server.go
@@ -21,7 +21,7 @@ type Server struct {
 	ASOnly         bool
 	PrefixOnly     bool
 	OtherCountries string
-	FileMaxsize    int
+	FileMaxsize    int64
 }
 
 type Servers []Server
